+++ NAMSC_lib/src/Novel/Data/Visual/Scenery/Scenery.cpp
@@ -77,9 +77,9 @@ bool Scenery::operator==(const Scenery& obj) const noexcept
 void Scenery::serializableLoad(QDataStream& dataStream)
 {
 	dataStream >> backgroundAssetImageName_ >> musicPlaylist;
-	uint displayedCharactersSize, displayedSceneryObjectsSize, soundsSize;
+	size_t displayedCharactersSize, displayedSceneryObjectsSize, soundsSize;
 	dataStream >> displayedCharactersSize >> displayedSceneryObjectsSize >> soundsSize;
-	for (uint i = 0; i != displayedCharactersSize; ++i)
+	for (size_t i = 0; i != displayedCharactersSize; ++i)
 	{
 		Character character;
 		dataStream >> character;
@@ -88,7 +88,7 @@ void Scenery::serializableLoad(QDataStream& dataStream)
 		displayedCharacters_[name] = std::move(character);
 	}
 
-	for (uint i = 0; i != displayedSceneryObjectsSize; ++i)
+	for (size_t i = 0; i != displayedSceneryObjectsSize; ++i)
 	{
 		SceneryObject sceneryObject;
 		dataStream >> sceneryObject;
@@ -96,7 +96,7 @@ void Scenery::serializableLoad(QDataStream& dataStream)
 		displayedSceneryObjects_[name] = std::move(sceneryObject);
 	}
 
-	for (uint i = 0; i != soundsSize; ++i)
+	for (size_t i = 0; i != soundsSize; ++i)
 	{
 		Sound sound;
 		dataStream >> sound;
